+++ src/main/java/gene/recombine/stuhubsys/vo/Student.java
@@ -14,9 +14,9 @@ public class Student extends Person {
     @Schema(description = "学生id")
     private String studentId;
     @Schema(description = "班级id")
-    private Integer classId;
+    private Long classId;
     @Schema(description = "专业id")
-    private Integer majorId;
+    private Long majorId;
     @Schema(description = "报名号")
     private String signUpCode;
     @Schema(description = "报名记录id列表")
